# Add support for netty native transports (epoll/kqueue)