https://github.com/netty/netty/wiki/Native-transports