@omarkilani 2019-04-02
Some performance numbers from a n1-highcpu-96 on GCP:

Java Version:

```
# java -version
openjdk version "12-testing" 2019-03-19
OpenJDK Runtime Environment (build 12-testing+0-builds.shipilev.net-openjdk-jdk12-b152-20190329-jdk-1233jdk-12-ga)
OpenJDK 64-Bit Server VM (build 12-testing+0-builds.shipilev.net-openjdk-jdk12-b152-20190329-jdk-1233jdk-12-ga, mixed mode, sharing)
```

JVM flags (simulate Remember The Milk environment):

```
java -XX:+UnlockExperimentalVMOptions -XX:+UseShenandoahGC \
       -XX:+ParallelRefProcEnabled \
       -Xss256k -Xmx70G \
       -XX:+AlwaysPreTouch \
       -XX:+UseTransparentHugePages \
       -XX:+ClassUnloadingWithConcurrentMark \
       -XX:+AggressiveOpts \
       -XX:+UseStringDeduplication \
       -XX:+UseNUMA \
       -XX:+UseCondCardMark \
       -XX:+ExitOnOutOfMemoryError \
        -jar $@
```

ab flags:

`ab -c 100 -n 10000000 -k http://localhost:8080/bounded-cycle-time-nonblocking`

Without native transport:

```
Server Software:
Server Hostname:        localhost
Server Port:            8080

Document Path:          /bounded-cycle-time-nonblocking
Document Length:        89 bytes

Concurrency Level:      100
Time taken for tests:   190.175 seconds
Complete requests:      10000000
Failed requests:        0
Keep-Alive requests:    10000000
Total transferred:      1920000000 bytes
HTML transferred:       890000000 bytes
Requests per second:    52583.04 [#/sec] (mean)
Time per request:       1.902 [ms] (mean)
Time per request:       0.019 [ms] (mean, across all concurrent requests)
Transfer rate:          9859.32 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       2
Processing:     0    2   1.9      2     673
Waiting:        0    2   1.9      2     673
Total:          0    2   1.9      2     674

Percentage of the requests served within a certain time (ms)
  50%      2
  66%      2
  75%      2
  80%      2
  90%      2
  95%      3
  98%      4
  99%      4
 100%    674 (longest request)
```

With native transport:

```

Server Software:
Server Hostname:        localhost
Server Port:            8080

Document Path:          /bounded-cycle-time-nonblocking
Document Length:        89 bytes

Concurrency Level:      100
Time taken for tests:   187.441 seconds
Complete requests:      10000000
Failed requests:        0
Keep-Alive requests:    10000000
Total transferred:      1920000000 bytes
HTML transferred:       890000000 bytes
Requests per second:    53350.23 [#/sec] (mean)
Time per request:       1.874 [ms] (mean)
Time per request:       0.019 [ms] (mean, across all concurrent requests)
Transfer rate:          10003.17 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       2
Processing:     0    2   0.5      2      30
Waiting:        0    2   0.5      2      30
Total:          0    2   0.5      2      30

Percentage of the requests served within a certain time (ms)
  50%      2
  66%      2
  75%      2
  80%      2
  90%      2
  95%      3
  98%      4
  99%      4
 100%     30 (longest request)
```