@omarkilani 2019-04-03
FWIW, using a multi-threaded testing tool (https://github.com/superfly/burn) the performance difference is huge:

Without native transport:

```
Burning http://localhost:8080/bounded-cycle-time-nonblocking for 120s

Statistical Analysis:
+-----------------------+--------+---------+---------+----------+-------+---------+-----------+----------+
|        Metric         |  p50   |   p75   |   p95   |   p99    |  Min  |  Mean   |    Max    | Std. Dev |
+-----------------------+--------+---------+---------+----------+-------+---------+-----------+----------+
| Connect               | 71µs   | 88µs    | 145µs   | 891µs    | 28µs  | 112µs   | 53.844ms  | 535µs    |
| Request fully written | 28µs   | 34µs    | 72µs    | 1.681ms  | 11µs  | 103µs   | 85.312ms  | 947µs    |
| Response fully read   | 3.36ms | 3.566ms | 5.305ms | 12.367ms | 442µs | 3.652ms | 752.092ms | 4.481ms  |
+-----------------------+--------+---------+---------+----------+-------+---------+-----------+----------+

Meta:
  Requests Count |         3624734  
  Time spent     | 1m59.999836381s  
  RPS            |    30206.157853  
  TLS Resumed    |               0  

Status code:
200 => 3624741

Errors:
0 errors.

```

With native transport:

```
Burning http://localhost:8080/bounded-cycle-time-nonblocking for 120s

Statistical Analysis:
+-----------------------+---------+---------+---------+---------+-------+---------+-----------+----------+
|        Metric         |   p50   |   p75   |   p95   |   p99   |  Min  |  Mean   |    Max    | Std. Dev |
+-----------------------+---------+---------+---------+---------+-------+---------+-----------+----------+
| Connect               | 66µs    | 81µs    | 137µs   | 562µs   | 26µs  | 92µs    | 74.269ms  | 326µs    |
| Request fully written | 26µs    | 32µs    | 53µs    | 589µs   | 10µs  | 79µs    | 115.482ms | 974µs    |
| Response fully read   | 2.002ms | 2.132ms | 2.989ms | 6.783ms | 316µs | 2.218ms | 163.84ms  | 1.8ms    |
+-----------------------+---------+---------+---------+---------+-------+---------+-----------+----------+

Meta:
  Requests Count |         6524323  
  Time spent     | 1m59.999471636s  
  RPS            |    54369.597724  
  TLS Resumed    |               0  

Status code:
200 => 6524334

Errors:
0 errors.
```